// src/model/ChipModel.ts
export enum CardStatus {
    PENDING = 'warning',
    TOPICKUP = 'secondary',
    PICKEDUP = 'primary',
    RETURNED = 'success'
}

export enum ConditionStatus {
    GOOD = 'success',
    MISPLACED = 'warning',
    DAMAGED = 'danger',
}

export enum AvailabilityStatus {
    AVAILABLE = 'success',
    NOT_AVAILABLE = 'danger',
}